+++ app/src/main/java/com/cnc/hcm/cnctrack/model/RemoveDeviceFromTaskResult.java
@@ -10,16 +10,16 @@ import com.google.gson.annotations.SerializedName;
 public class RemoveDeviceFromTaskResult {
     @SerializedName("statusCode")
     @Expose
-    private Long statusCode;
+    private int statusCode;
     @SerializedName("message")
     @Expose
     private String message;
 
-    public Long getStatusCode() {
+    public int getStatusCode() {
         return statusCode;
     }
 
-    public void setStatusCode(Long statusCode) {
+    public void setStatusCode(int statusCode) {
         this.statusCode = statusCode;
     }
